requirements.txt
python-telegram-bot==20.3
beautifulsoup4==4.12.2
requests==2.31.0
nest_asyncio==1.5.6
